Add storyboard compatibility, declare class public
Storyboards use init(decoder) to initialize their content, so they avoid your convenience inits. This leads setting an action to the button to be useless because the target-action isn't set. I've changed action property setter to solve this. Calling addTarget() multiple times without removeTarget() shouldn't be a problem, because all targets are stored in a dictionary (so setting new value basically overrides old).